Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    double myPow(double x, long long n) {
+    double myPow(double x, unsigned int n) {
         if (n == 0) return 1;          
         if (n < 0) return (1 / myPow(x, -n));
